Simplify: extract personal-sync group from api.js (slice 2)
Move the 14 personal-sync methods (playlists / history / resume / settings /
provisionPersonalEncryption) into src/api/personal.js as
createPersonalApi({ ctx }), spread into the api object.

Even cleaner than slice 1: the group's own banner documents that these are
called via the shared HRPC handler table as api.method(request) with `this`
unbound, so they never reference `this` — only the injected ctx. Closure-var
audit confirms ctx is the sole dependency (no this.*, no other closure leaks).
api.js 4742 -> 4651 lines.
