+++ packages/backend/src/api.js
@@ -28,6 +28,7 @@ import { collectCorestoreGarbage } from './corestore-gc.js'
 import { peerHasFullRange, collectFullCopyPeers, assessOffloadEligibility } from './upload-offload.js'
 import { verifySignedChannelRootDescriptor } from './channel-descriptor.js'
 import { createCommentsApi } from './api/comments.js'
+import { createPersonalApi } from './api/personal.js'
 
 /**
  * @typedef {import('./types.js').StorageContext} StorageContext
@@ -2411,98 +2412,7 @@ export function createApi({
       return enriched;
     },
 
-    // ============================================
-    // Personal Sync: Playlists / History / Settings
-    // (private per-identity multi-writer store, synced across the user's devices)
-    //
-    // These are registered through the shared HRPC handler table, which calls
-    // them as `api.method(request)` with `this` unbound — so they take the
-    // decoded request object and return the response envelope directly, and
-    // must not rely on `this`.
-    // ============================================
-
-    async getPlaylists() {
-      if (!ctx.personal) return { playlists: [] };
-      return { playlists: await ctx.personal.listPlaylists() };
-    },
-    async getPlaylistItems(req = {}) {
-      if (!ctx.personal) return { items: [] };
-      return { items: await ctx.personal.listPlaylistItems(req.playlistId) };
-    },
-    async createPlaylist(req = {}) {
-      if (!ctx.personal?.writable) throw new Error('No writable personal store (create or activate an identity first)');
-      const id = await ctx.personal.createPlaylist({ name: req.name || '', description: req.description || '' });
-      return { success: true, id };
-    },
-    async updatePlaylist(req = {}) {
-      if (!ctx.personal?.writable) throw new Error('No writable personal store');
-      await ctx.personal.updatePlaylist(req.id, { name: req.name, description: req.description });
-      return { success: true };
-    },
-    async deletePlaylist(req = {}) {
-      if (!ctx.personal?.writable) throw new Error('No writable personal store');
-      await ctx.personal.deletePlaylist(req.id);
-      return { success: true };
-    },
-    async addToPlaylist(req = {}) {
-      if (!ctx.personal?.writable) throw new Error('No writable personal store');
-      await ctx.personal.addToPlaylist(req.playlistId, { channelKey: req.channelKey, videoId: req.videoId, videoKey: req.videoKey });
-      return { success: true };
-    },
-    async removeFromPlaylist(req = {}) {
-      if (!ctx.personal?.writable) throw new Error('No writable personal store');
-      await ctx.personal.removeFromPlaylist(req.playlistId, req.videoKey);
-      return { success: true };
-    },
-
-    async logWatchHistory(req = {}) {
-      if (!ctx.personal?.writable) throw new Error('No writable personal store');
-      const eventId = await ctx.personal.logHistory(req);
-      return { success: true, eventId };
-    },
-    async getWatchHistory(req = {}) {
-      if (!ctx.personal) return { entries: [] };
-      return { entries: await ctx.personal.listHistory({ limit: req.limit || 100 }) };
-    },
-    async getResumePosition(req = {}) {
-      if (!ctx.personal) return { found: false };
-      const resume = await ctx.personal.getResume(req.videoKey);
-      return resume ? { found: true, resume } : { found: false };
-    },
-    async listResumePositions() {
-      if (!ctx.personal) return { entries: [] };
-      return { entries: await ctx.personal.listResume() };
-    },
-
-    async setPersonalSetting(req = {}) {
-      if (!ctx.personal?.writable) throw new Error('No writable personal store');
-      // Values arrive JSON-encoded over HRPC so a setting can hold any JSON type.
-      let value = req.value;
-      if (typeof value === 'string') {
-        try { value = JSON.parse(value); } catch { /* keep raw string */ }
-      }
-      await ctx.personal.setSetting(req.key, value);
-      return { success: true };
-    },
-    async getPersonalSettings() {
-      if (!ctx.personal) return { settings: [] };
-      const settings = await ctx.personal.getSettings();
-      return { settings: Object.entries(settings).map(([key, value]) => ({ key, value: JSON.stringify(value) })) };
-    },
-
-    /**
-     * Provision the at-rest encryption secret (from the device's native
-     * keychain) for the active identity's personal store, opening it encrypted.
-     * The platform reads-or-generates the secret in the keychain and passes it
-     * here; when omitted, the backend generates one and returns it so the
-     * platform can persist it to the keychain. Must be called before the store
-     * first opens to take effect.
-     */
-    async provisionPersonalEncryption(req = {}) {
-      if (!ctx.personalManager) return { success: false, error: 'personal store unavailable' };
-      const result = await ctx.personalManager.provisionSecret({ secret: req.secret || undefined });
-      return { success: !!result.success, secret: result.secret, encrypted: !!result.encrypted, error: result.error };
-    },
+    ...createPersonalApi({ ctx }),
 
     /**
      * Resolve the signed channel root descriptor for a locally available
